requirements.txt
Flask==3.0.0
Flask-CORS==4.0.0
openai==1.12.0
PyPDF2==3.0.1
Pillow==10.2.0
opencv-python-headless==4.9.0.80
python-dotenv==1.0.0
Werkzeug==3.0.1
gunicorn==21.2.0